requirements.txt
requests==2.28.2
pymongo==4.3.3
python-dotenv==0.21.1